+++ src/AeFinder.App/AeFinder.App.csproj
@@ -8,13 +8,11 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="AElf.CSharp.Core" Version="1.6.0" />
         <PackageReference Include="GraphQL.Server.Transports.AspNetCore" Version="7.1.1" />
         <PackageReference Include="GraphQL.Server.Ui.GraphiQL" Version="7.1.1" />
         <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
         <PackageReference Include="Volo.Abp.AspNetCore.Mvc" Version="8.1.4" />
         <PackageReference Include="GraphQL.Server.Ui.Playground" Version="7.1.1" />
-        <PackageReference Include="AElf.Client" Version="1.2.1" />
         <PackageReference Include="OpenTelemetry.Extensions.Hosting" Version="1.9.0" />
         <PackageReference Include="OpenTelemetry.Instrumentation.AspNetCore" Version="1.9.0" />
         <PackageReference Include="OpenTelemetry.Exporter.Prometheus.AspNetCore" Version="1.8.0-rc.1" />
