+++ src/AeFinder.Application.Contracts/AeFinder.Application.Contracts.csproj
@@ -14,6 +14,14 @@
   </ItemGroup>
 
   <ItemGroup>
+    <PackageReference Include="AElf.Client" Version="1.3.0" />
+    <PackageReference Include="AElf.Core" Version="1.6.0" />
+    <PackageReference Include="AElf.Cryptography" Version="1.6.0" />
+    <PackageReference Include="AElf.CSharp.Core" Version="1.6.0" />
+    <PackageReference Include="AElf.Kernel.Types" Version="1.6.0" />
+    <PackageReference Include="AElf.Types" Version="1.6.0" />
+    <PackageReference Include="GraphQL.Client" Version="6.1.0" />
+    <PackageReference Include="GraphQL.Client.Serializer.Newtonsoft" Version="6.1.0" />
     <PackageReference Include="Com.Ctrip.Framework.Apollo.Configuration" Version="2.10.2" />
     <PackageReference Include="Microsoft.Orleans.Sdk" Version="8.1.0" />
     <PackageReference Include="Volo.Abp.AspNetCore.Serilog" Version="8.1.4" />
@@ -25,6 +33,10 @@
     <PackageReference Include="Volo.Abp.FeatureManagement.Application.Contracts" Version="8.1.4" />
     <PackageReference Include="Volo.Abp.SettingManagement.Application.Contracts" Version="8.1.4" />
     <PackageReference Include="Volo.Abp.TenantManagement.Application.Contracts" Version="8.1.4" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Folder Include="Contract\" />
   </ItemGroup>
