+++ src/AeFinder.Application.Contracts/Contract/CaContract.c.cs
