+++ src/AeFinder.Application.Contracts/Contract/CaContractImpl.c.cs
