+++ src/AeFinder.Application/AeFinder.Application.csproj
@@ -22,7 +22,6 @@
   <ItemGroup>
     <PackageReference Include="AWSSDK.S3" Version="3.7.401.2" />
     <PackageReference Include="DotNetZip" Version="1.16.0" />
-    <PackageReference Include="GraphQL.Client" Version="6.0.3" />
     <PackageReference Include="IdentityModel.OidcClient" Version="5.2.1" />
     <PackageReference Include="NUglify" Version="1.21.8" />
     <PackageReference Include="Remotion.Linq" Version="2.2.0" />
