+++ src/AeFinder.AuthServer/AeFinder.AuthServer.csproj
@@ -32,12 +32,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="AElf.Client" Version="1.2.3" />
-    <PackageReference Include="AElf.Core" Version="1.6.0" />
-    <PackageReference Include="AElf.Cryptography" Version="1.6.0" />
-    <PackageReference Include="AElf.CSharp.Core" Version="1.6.0" />
-    <PackageReference Include="AElf.Kernel.Types" Version="1.6.0" />
-    <PackageReference Include="AElf.Types" Version="1.6.0" />
     <PackageReference Include="GraphQL.Client" Version="6.1.0" />
     <PackageReference Include="GraphQL.Client.Serializer.Newtonsoft" Version="6.1.0" />
     <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="4.10.0" />
@@ -64,10 +58,6 @@
       <GlobalPropertiesToRemove></GlobalPropertiesToRemove>
     </ProjectReference>
     <ProjectReference Include="..\AeFinder.MongoDB\AeFinder.MongoDB.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="Contract\" />
   </ItemGroup>
 
 </Project>
