--- src/AeFinder.AuthServer/Options/ChainOptions.cs
@@ -1,8 +1,0 @@
-using System.Collections.Generic;
-
-namespace AeFinder.Options;
-
-public class ChainOptions
-{
-    public Dictionary<string, ChainInfo> ChainInfos { get; set; }
-}
