+++ src/AeFinder.AuthServer/SignatureGrantHandler.cs
@@ -39,11 +39,7 @@ public class SignatureGrantHandler: ITokenExtensionGrant, ITransientDependency
     private ILogger<SignatureGrantHandler> _logger;
     private IAbpDistributedLock _distributedLock;
     private const string LockKeyPrefix = "AeFinder:Auth:SignatureGrantHandler:";
-    private const string GetHolderInfoMethodName = "GetHolderInfo";
-    private const string PortKeyAppId = "PortKey";
-    private const string NightElfAppId = "NightElf";
-    private const string CrossChainContractName = "AElf.ContractNames.CrossChain";
-    // private IClusterClient _clusterClient;
+    
     private SignatureOptions _signatureOptions;
     private ChainOptions _chainOptions;
     private IUserInformationProvider _userInformationProvider;
@@ -58,7 +54,6 @@ public class SignatureGrantHandler: ITokenExtensionGrant, ITransientDependency
         var caHash = context.Request.GetParameter("ca_hash").ToString();
         var timestampVal = context.Request.GetParameter("timestamp").ToString();
         var address = context.Request.GetParameter("address").ToString();
-        var source = context.Request.GetParameter("source").ToString();
         var userName = context.Request.GetParameter("uname").ToString();
 
         //Before opening registration, must first log in with a regular account
@@ -96,33 +91,31 @@ public class SignatureGrantHandler: ITokenExtensionGrant, ITransientDependency
                 $"The time should be {timeRange} minutes before and after the current time.");
         }
         
+        _userInformationProvider = context.HttpContext.RequestServices.GetRequiredService<IUserInformationProvider>();
         _logger = context.HttpContext.RequestServices.GetRequiredService<ILogger<SignatureGrantHandler>>();
         //Validate public key and signature
-        var newSignText = """
-                          Welcome to AeFinder! Click to sign in to the AeFinder platform! This request will not trigger any blockchain transaction or cost any gas fees.
-
-                          signature: 
-                          """+string.Join("-", address, timestampVal);
-        _logger.LogInformation("newSignText:{newSignText}",newSignText);
-        if (!CryptoHelper.RecoverPublicKey(signature, HashHelper.ComputeFrom(Encoding.UTF8.GetBytes(newSignText).ToHex()).ToByteArray(),
-                out var managerPublicKey))
+//         var newSignText = """
+//                           Welcome to AeFinder! Click to sign in to the AeFinder platform! This request will not trigger any blockchain transaction or cost any gas fees.
+//
+//                           signature: 
+//                           """+string.Join("-", address, timestampVal);
+//         _logger.LogInformation("newSignText:{newSignText}",newSignText);
+        if (!_userInformationProvider.RecoverPublicKey(address, timestampVal, signature, out var managerPublicKey))
         {
             return GetForbidResult(OpenIddictConstants.Errors.InvalidRequest, "Signature validation failed new.");
         }
-        
-        if (!CryptoHelper.RecoverPublicKey(signature, HashHelper.ComputeFrom(string.Join("-", address, timestampVal)).ToByteArray(),
-                out var managerPublicKeyOld))
+
+        if (!_userInformationProvider.RecoverPublicKeyOld(address, timestampVal, signature, out var managerPublicKeyOld))
         {
             return GetForbidResult(OpenIddictConstants.Errors.InvalidRequest, "Signature validation failed old.");
         }
 
-        if (!(managerPublicKey.ToHex() == publicKeyVal || managerPublicKeyOld.ToHex() == publicKeyVal))
+        if (!_userInformationProvider.CheckPublicKey(managerPublicKey, managerPublicKeyOld, publicKeyVal))
         {
             return GetForbidResult(OpenIddictConstants.Errors.InvalidRequest, "Invalid publicKey or signature.");
         }
-        
+
         _distributedLock = context.HttpContext.RequestServices.GetRequiredService<IAbpDistributedLock>();
-        // _clusterClient = context.HttpContext.RequestServices.GetRequiredService<IClusterClient>();
         _chainOptions = context.HttpContext.RequestServices.GetRequiredService<IOptionsMonitor<ChainOptions>>().CurrentValue;
         _logger.LogInformation(
             "publicKeyVal:{0}, signatureVal:{1}, address:{2}, caHash:{3}, chainId:{4}, timestamp:{5}",
@@ -130,7 +123,7 @@ public class SignatureGrantHandler: ITokenExtensionGrant, ITransientDependency
 
         
         var userManager = context.HttpContext.RequestServices.GetRequiredService<IdentityUserManager>();
-        _userInformationProvider = context.HttpContext.RequestServices.GetRequiredService<IUserInformationProvider>();
+        
         //Before opening registration, must first log in with a regular account
         var user = await userManager.FindByNameAsync(userName);
         if (user == null || user.IsDeleted)
@@ -160,8 +153,7 @@ public class SignatureGrantHandler: ITokenExtensionGrant, ITransientDependency
                 return GetForbidResult(OpenIddictConstants.Errors.InvalidRequest, "User has already linked another Portkey wallet address.");
             }
 
-            var managerCheck = await CheckAddressAsync(chainId, _signatureOptions.PortkeyV2GraphQLUrl, caHash, signAddress,
-                _chainOptions);
+            var managerCheck = await _userInformationProvider.CheckAddressAsync(chainId, _signatureOptions.PortkeyV2GraphQLUrl, caHash, signAddress);
             if (!managerCheck.HasValue || !managerCheck.Value)
             {
                 _logger.LogError("Manager validation failed. caHash:{0}, address:{1}, chainId:{2}",
@@ -169,7 +161,7 @@ public class SignatureGrantHandler: ITokenExtensionGrant, ITransientDependency
                 return GetForbidResult(OpenIddictConstants.Errors.InvalidRequest, "Manager validation failed.");
             }
 
-            addressInfos = await GetAddressInfosAsync(caHash);
+            addressInfos = await _userInformationProvider.GetAddressInfosAsync(caHash);
             userExtensionDto.CaAddressList = addressInfos;
         }
         else
@@ -305,158 +297,6 @@ public class SignatureGrantHandler: ITokenExtensionGrant, ITransientDependency
         }
 
         return resources;
-    }
-    
-    
-    private async Task<bool?> CheckAddressAsync(string chainId, string graphQlUrl, string caHash, string manager,
-        ChainOptions chainOptions)
-    {
-        var graphQlResult = await CheckAddressFromGraphQlAsync(graphQlUrl, caHash, manager);
-        if (!graphQlResult.HasValue || !graphQlResult.Value)
-        {
-            _logger.LogDebug("graphql is invalid.");
-            return await CheckAddressFromContractAsync(chainId, caHash, manager, chainOptions);
-        }
-
-        return true;
-    }
-    
-    private async Task<bool?> CheckAddressFromContractAsync(string chainId, string caHash, string manager,
-        ChainOptions chainOptions)
-    {
-        var param = new GetHolderInfoInput
-        {
-            CaHash = Hash.LoadFromHex(caHash),
-            LoginGuardianIdentifierHash = Hash.Empty
-        };
-
-        var output =
-            await CallTransactionAsync<GetHolderInfoOutput>(chainId, GetHolderInfoMethodName, param, false,
-                chainOptions);
-
-        return output?.ManagerInfos?.Any(t => t.Address.ToBase58() == manager);
-    }
-    
-    private async Task<bool?> CheckAddressFromGraphQlAsync(string url, string caHash,
-        string managerAddress)
-    {
-        var cHolderInfos = await GetHolderInfosAsync(url, caHash);
-        var caHolder = cHolderInfos?.CaHolderInfo?.SelectMany(t => t.ManagerInfos);
-        return caHolder?.Any(t => t.Address == managerAddress);
-    }
-    
-    private async Task<T> CallTransactionAsync<T>(string chainId, string methodName, IMessage param,
-        bool isCrossChain, ChainOptions chainOptions) where T : class, IMessage<T>, new()
-    {
-        try
-        {
-            var chainInfo = chainOptions.ChainInfos[chainId];
-
-            var client = new AElfClient(chainInfo.AElfNodeBaseUrl);
-            await client.IsConnectedAsync();
-            var address = client.GetAddressFromPrivateKey(_signatureOptions.CommonPrivateKeyForCallTx);
-
-            var contractAddress = isCrossChain
-                ? (await client.GetContractAddressByNameAsync(HashHelper.ComputeFrom(CrossChainContractName)))
-                .ToBase58()
-                : chainInfo.CAContractAddress;
-
-            var transaction =
-                await client.GenerateTransactionAsync(address, contractAddress,
-                    methodName, param);
-
-            var txWithSign = client.SignTransaction(_signatureOptions.CommonPrivateKeyForCallTx, transaction);
-            var result = await client.ExecuteTransactionAsync(new ExecuteTransactionDto
-            {
-                RawTransaction = txWithSign.ToByteArray().ToHex()
-            });
-
-            var value = new T();
-            value.MergeFrom(ByteArrayHelper.HexStringToByteArray(result));
-            return value;
-        }
-        catch (Exception e)
-        {
-            if (methodName != GetHolderInfoMethodName)
-            {
-                _logger.LogError(e, "CallTransaction error, chain id:{chainId}, methodName:{methodName}", chainId,
-                    methodName);
-            }
-
-            _logger.LogError(e, "CallTransaction error, chain id:{chainId}, methodName:{methodName}", chainId,
-                methodName);
-            return null;
-        }
-    }
-    
-    private async Task<HolderInfoIndexerDto> GetHolderInfosAsync(string url, string caHash)
-    {
-        using var graphQlClient = new GraphQLHttpClient(url, new NewtonsoftJsonSerializer());
-        var request = new GraphQLRequest
-        {
-            Query = @"
-			    query($caHash:String,$skipCount:Int!,$maxResultCount:Int!) {
-                    caHolderInfo(dto: {caHash:$caHash,skipCount:$skipCount,maxResultCount:$maxResultCount}){
-                            id,chainId,caHash,caAddress,originChainId,managerInfos{address,extraData}}
-                }",
-            Variables = new
-            {
-                caHash, skipCount = 0, maxResultCount = 10
-            }
-        };
-
-        var graphQlResponse = await graphQlClient.SendQueryAsync<HolderInfoIndexerDto>(request);
-        return graphQlResponse.Data;
-    }
-    
-    private async Task<List<UserChainAddressDto>> GetAddressInfosAsync(string caHash)
-    {
-        var addressInfos = new List<UserChainAddressDto>();
-        var holderInfoDto = await GetHolderInfosAsync(_signatureOptions.PortkeyV2GraphQLUrl, caHash);
-
-        var chainIds = new List<string>();
-        if (holderInfoDto != null && !holderInfoDto.CaHolderInfo.IsNullOrEmpty())
-        {
-            addressInfos.AddRange(holderInfoDto.CaHolderInfo.Select(t => new UserChainAddressDto
-                { ChainId = t.ChainId, Address = t.CaAddress }));
-            chainIds = holderInfoDto.CaHolderInfo.Select(t => t.ChainId).ToList();
-        }
-
-        var chains = _chainOptions.ChainInfos.Select(key => _chainOptions.ChainInfos[key.Key])
-            .Select(chainOptionsChainInfo => chainOptionsChainInfo.ChainId).Where(t => !chainIds.Contains(t));
-
-        foreach (var chainId in chains)
-        {
-            try
-            {
-                var addressInfo = await GetAddressInfoAsync(chainId, caHash);
-                addressInfos.Add(addressInfo);
-            }
-            catch (Exception e)
-            {
-                _logger.LogError(e, "get holder from chain error, caHash:{caHash}", caHash);
-            }
-        }
-
-        return addressInfos;
-    }
-    
-    private async Task<UserChainAddressDto> GetAddressInfoAsync(string chainId, string caHash)
-    {
-        var param = new GetHolderInfoInput
-        {
-            CaHash = Hash.LoadFromHex(caHash),
-            LoginGuardianIdentifierHash = Hash.Empty
-        };
-
-        var output = await CallTransactionAsync<GetHolderInfoOutput>(chainId, GetHolderInfoMethodName, param, false,
-            _chainOptions);
-
-        return new UserChainAddressDto()
-        {
-            Address = output.CaAddress.ToBase58(),
-            ChainId = chainId
-        };
     }
 
     private async Task<bool> CreateUserAsync(
