+++ src/AeFinder.Domain.Shared/Options/ChainOptions.cs
@@ -1,4 +1,11 @@
+using System.Collections.Generic;
+
 namespace AeFinder.Options;
+
+public class ChainOptions
+{
+    public Dictionary<string, ChainInfo> ChainInfos { get; set; }
+}
 
 public class ChainInfo
 {
