+++ src/AeFinder.Domain.Shared/Options/SignatureOptions.cs
